+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Change Log #
 
-## Version 1.10 (stable-19104)
+## Version 1.10 (develop-19144)
 
-* Support lock as trigger device
+* Add trigger quieting--a span of time after which device are turned off (manually) that the triggers are ignored (so if one turns off lights manually in a room with a motion sensor, the motion of exiting the room does not turn the lights immediately back on).
+* Support lock as trigger device.
 * Fix bug that causes errorneous notification of modified config although no changes have been made.
 
 ## Version 1.9 (released)